backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS & middleware
python-dotenv==1.0.0

# File handling
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: Email sending (for future features)
# aiosmtplib==2.0.2

# Optional: Redis for caching (for future features)
# redis==5.0.1
